+++ src/Components/ContentElement/PanelStartElementController.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ContaoBootstrap\Panel\Components\ContentElement;
+
+use Contao\CoreBundle\ServiceAnnotation\ContentElement;
+
+/** @ContentElement("bs_panel_start", category="bs_panel") */
+final class PanelStartElementController extends AbstractPanelStartElementController
+{
+}