+++ src/ContaoBootstrapPanelBundle.php
@@ -1,28 +1,11 @@
 <?php
 
-/**
- * Contao Bootstrap panel.
- *
- * @package    contao-bootstrap
- * @subpackage Panel
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2020 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0-or-later https://github.com/contao-bootstrap/panel/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Panel;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-/**
- * Class ContaoBootstrapPanelBundle.
- *
- * @package ContaoBootstrap\Panel
- */
-class ContaoBootstrapPanelBundle extends Bundle
+final class ContaoBootstrapPanelBundle extends Bundle
 {
-
 }