+++ src/Resources/contao/config/config.php
@@ -1,34 +1,6 @@
 <?php
 
-/**
- * Contao Bootstrap panel.
- *
- * @package    contao-bootstrap
- * @subpackage Panel
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2020 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0-or-later https://github.com/contao-bootstrap/panel/blob/master/LICENSE
- * @filesource
- */
-
 declare(strict_types=1);
-
-use ContaoBootstrap\Panel\Components\ContentElement\PanelGroupStartElement;
-use ContaoBootstrap\Panel\Components\ContentElement\PanelGroupEndElement;
-use ContaoBootstrap\Panel\Components\ContentElement\PanelSingleElement;
-use ContaoBootstrap\Panel\Components\ContentElement\PanelStartElement;
-use ContaoBootstrap\Panel\Components\ContentElement\PanelEndElement;
-
-/*
- * Content elements
- */
-
-$GLOBALS['TL_CTE']['bs_panel']['bs_panel_group_start'] = PanelGroupStartElement::class;
-$GLOBALS['TL_CTE']['bs_panel']['bs_panel_group_end']   = PanelGroupEndElement::class;
-$GLOBALS['TL_CTE']['bs_panel']['bs_panel_single']      = PanelSingleElement::class;
-$GLOBALS['TL_CTE']['bs_panel']['bs_panel_start']       = PanelStartElement::class;
-$GLOBALS['TL_CTE']['bs_panel']['bs_panel_end']         = PanelEndElement::class;
-
 
 /*
  * Wrappers