+++ src/Resources/contao/dca/tl_content.php
@@ -1,15 +1,6 @@
 <?php
 
-/**
- * Contao Bootstrap panel.
- *
- * @package    contao-bootstrap
- * @subpackage Panel
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2020 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0-or-later https://github.com/contao-bootstrap/panel/blob/master/LICENSE
- * @filesource
- */
+declare(strict_types=1);
 
 /*
  * Palettes
@@ -65,17 +56,15 @@ $GLOBALS['TL_DCA']['tl_content']['metapalettes']['bs_panel_single'] = [
  * Fields
  */
 
-$GLOBALS['TL_DCA']['tl_content']['fields']['bs_expanded'] = array(
-    'label'     => &$GLOBALS['TL_LANG']['tl_content']['bs_expanded'],
+$GLOBALS['TL_DCA']['tl_content']['fields']['bs_expanded'] = [
     'exclude'   => true,
     'inputType' => 'checkbox',
     'default'   => false,
-    'eval'      => array('tl_class' => 'w50 m12'),
+    'eval'      => ['tl_class' => 'w50 m12'],
     'sql'       => "char(1) NOT NULL default ''",
-);
+];
 
 $GLOBALS['TL_DCA']['tl_content']['fields']['bs_panel_name'] = [
-    'label'         => &$GLOBALS['TL_LANG']['tl_content']['bs_panel_name'],
     'exclude'       => true,
     'inputType'     => 'text',
     'reference'     => &$GLOBALS['TL_LANG']['tl_content'],
@@ -83,9 +72,6 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['bs_panel_name'] = [
         'includeBlankOption' => true,
         'chosen'             => true,
         'tl_class'           => 'w50',
-    ],
-    'save_callback' => [
-        ['contao_bootstrap.panel.listener.dca.content', 'generatePanelName'],
     ],
     'sql'           => "varchar(64) NOT NULL default ''",
 ];